Build basic drawing function
Added an on click event that changes the background color property
of a '.square' div when the mouse is over it. You dont need the hover
event to fire to fire this, but eventually, the  hover event will
contain a sort of 'preview' mode, (i.e. change the square color
temporarily)